arm: spl: normalize BOOT_DEVICE_NONE value
In order to support a boot device list in SPL, it is necessary to
be able to designate a value for "no entry" to serve as a list
terminator.

BOOT_DEVICE_NONE is a natural candidate, but it is not defined for
all architectures. In itself, that is not a problem, since it can be
conditionally defined in spl common code where it shall be needed.
However, BOOT_DEVICE_NONE value needs to be normalized across architectures,
otherwise the definition done in spl common code will differ from
other spl code, and weird bugs may happen.

The standard value for BOOT_DEVICE_NONE across most of U-Boot is 0.
Make it consistent by removing the only exception.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>